Bug 577431,  Skip files when dsymutil doesn't produce a .dSYM, r=josh, a=blocking2.0
UltraBlame original commit: ce8c5b97125cb9c22ef88b2c66cfb68bea9b9f9d

+++ toolkit/crashreporter/tools/symbolstore.py
@@ -710,6 +710,9 @@ class Dumper_Mac(Dumper):
         
         os.system("dsymutil %s %s >/dev/null" % (' '.join([a.replace('-a ', '--arch=') for a in self.archs]),
                                       file))
+        if not os.path.exists(dsymbundle):
+            
+            return False
         res = Dumper.ProcessFile(self, dsymbundle)
         
         shutil.rmtree(dsymbundle)
